requirements.txt
asgiref==3.7.2
Django==4.2.7
django-debug-toolbar==4.2.0
django-jazzmin==2.6.0
django-widget-tweaks==1.5.0
djangorestframework==3.14.0
fontawesomefree==6.4.2
Pillow==10.1.0
python-environ==0.4.54
pytz==2023.3.post1
sqlparse==0.4.4
tomlkit==0.12.2
tzdata==2023.3
